Refactor nearly everything.
Renamed a lot of functions and variables to be more descriptive
and adhere to pep8. Reordered many functions to both group related
functions and to follow the order in which these functions are
generally called. Refactored nearly every function in the use
code path to simplify their functionality. Made all of the data
structures read from the config parser object match each other
(lists of key/value pair tuples), as well as coerce the aliases
read from the stdIn to also match this format. Updated the
write_history function to write out more structured data. And so
much more. I should be able to do more atomic commits from this
point forward. Still is not a completely functional system yet.