Added Update Latest feature
Added a feature where the root user can run useupdate from the command
line and a symlink is created for each use package, but with the string
"latest" replacing the version number. Allows the end user to simply
run use <appname.-latest and they will get the latest installed version.